requirements2.txt
boto3==1.15.18
celery==5.0.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-sesame==2.1
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
oauthlib==3.1.0
Pillow==8.0.0
pipenv==2018.5.18
psycopg2-binary==2.8.6
python-decouple==3.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
whitenoise==5.2.0
